+++ lib/i18n/strings.g.dart
@@ -4,9 +4,9 @@
 /// To regenerate, run: `dart run slang`
 ///
 /// Locales: 1
-/// Strings: 896
+/// Strings: 899
 ///
-/// Built on 2026-06-18 at 16:10 UTC
+/// Built on 2026-06-19 at 15:52 UTC
 
 // coverage:ignore-file
 // ignore_for_file: type=lint, unused_import